requirements.txt
fastapi>=0.104.0
mangum>=0.17.0
boto3>=1.34.0
uvicorn>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
